+++ tests/test_suite.py
@@ -29,6 +29,9 @@ def load_test_genomes():
 def load_broken_genome():
     return glob.glob(os.path.join(str(os.getcwd()),'../tests/test_backend/other/broken_genome.fasta'))[0]
 
+def load_zipped_genome():
+    return glob.glob(os.path.join(str(os.getcwd()),'../tests/test_backend/zipped/Enterococcus_casseliflavus_EC20.fasta.gz'))[0]
+
 def load_test_wd_loc():
     loc = os.path.join(str(os.getcwd()),'../tests/test_backend/ecoli_wd')
     return loc
@@ -367,10 +370,12 @@ class VerifyBonus():
         self.genomes = load_test_genomes()
         self.wd_loc = load_test_wd_loc()
         self.s_wd_loc = load_solutions_wd()
+        self.zipped_genome = load_zipped_genome()
 
         #logging.shutdown()
         if os.path.isdir(self.wd_loc):
             shutil.rmtree(self.wd_loc)
+        os.mkdir(self.wd_loc)
 
     def run(self):
         self.setUp()
@@ -381,9 +386,29 @@ class VerifyBonus():
         self.test_drep_scaffold_level2()
         self.tearDown()
 
+        self.setUp()
+        self.test_drep_scaffold_level3()
+        self.tearDown()
+
+        self.setUp()
+        self.test_drep_scaffold_level4()
+        self.tearDown()
+
+        self.setUp()
+        self.test_drep_scaffold_level5()
+        self.tearDown()
+
+        self.setUp()
+        self.test_parse_stb()
+        self.tearDown()
+
+        self.setUp()
+        self.test_parse_stb2()
+        self.tearDown()
+
     def test_drep_scaffold_level(self):
         '''
-        test drep.d_filter.chdb_to_genomeInfo
+        test ScaffoldLevel_dRep.py basicaly
         '''
         script_loc = os.path.join(str(os.getcwd()),'../helper_scripts/ScaffoldLevel_dRep.py')
 
@@ -392,14 +417,14 @@ class VerifyBonus():
         subprocess.call(cmd, shell=True)
 
         outs = glob.glob(self.wd_loc + '/*')
-        print(outs)
+        assert(len(outs) == 3)
         f = [o for o in outs if 'DereplicationInfo.csv' in o][0]
         Rdb = pd.read_csv(f)
         assert len(Rdb) == 5
 
     def test_drep_scaffold_level2(self):
         '''
-        test drep.d_filter.chdb_to_genomeInfo
+        test ScaffoldLevel_dRep.py with --IgnoreSameScaffolds
         '''
         script_loc = os.path.join(str(os.getcwd()),'../helper_scripts/ScaffoldLevel_dRep.py')
 
@@ -408,29 +433,100 @@ class VerifyBonus():
         subprocess.call(cmd, shell=True)
 
         outs = glob.glob(self.wd_loc + '/*')
-        print(outs)
+        assert(len(outs) == 3)
         f = [o for o in outs if 'DereplicationInfo.csv' in o][0]
         Rdb = pd.read_csv(f)
         assert len(Rdb) == 0
 
+    def test_drep_scaffold_level3(self):
+        '''
+        test ScaffoldLevel_dRep.py with a single fasta file
+        '''
+        script_loc = os.path.join(str(os.getcwd()),'../helper_scripts/ScaffoldLevel_dRep.py')
 
-        # genomes = self.genomes
-        # workDirectory = drep.WorkDirectory.WorkDirectory(self.s_wd_loc)
-        #
-        # # Load chdb
-        # chdb = workDirectory.get_db('Chdb')
-        # # Make bdb
-        # bdb = drep.d_cluster.load_genomes(genomes)
-        # # Make Gdb
-        # Gdb = drep.d_filter.calc_genome_info(genomes)
-        #
-        # # Run
-        # Idb = drep.d_filter.chdb_to_genomeInfo(chdb)
-        # Tdb = drep.d_filter._validate_genomeInfo(Idb, bdb)
-        # Tdb = drep.d_filter._add_lengthN50(Tdb, bdb)
-        # t = Tdb[Tdb['genome'] == 'Enterococcus_casseliflavus_EC20.fasta']
-        # assert t['completeness'].tolist()[0] == 98.28
-        # assert t['length'].tolist()[0] == 3427276
+        cmd = "{0} -f {1} -o {2}".format(script_loc, self.genomes[0], self.wd_loc)
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+
+        outs = glob.glob(self.wd_loc + '/*')
+        assert(len(outs) == 4)
+        f = [o for o in outs if 'DereplicationInfo.csv' in o][0]
+        Rdb = pd.read_csv(f)
+        assert len(Rdb) == 0
+
+    def test_drep_scaffold_level4(self):
+        '''
+        test for graceful crash
+        '''
+        script_loc = os.path.join(str(os.getcwd()),'../helper_scripts/ScaffoldLevel_dRep.py')
+
+        cmd = "{0} -f {1} -o {2}".format(script_loc, self.zipped_genome, self.wd_loc)
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+
+        outs = glob.glob(self.wd_loc + '/*')
+        assert(len(outs) == 1), outs
+
+    def test_drep_scaffold_level5(self):
+        '''
+        test breaking things into chunks
+        '''
+        script_loc = os.path.join(str(os.getcwd()),'../helper_scripts/ScaffoldLevel_dRep.py')
+
+        cmd = "{0} -f {1} -o {2} -m 1000000".format(script_loc, [g for g in self.genomes if 'Enterococcus_faecalis_YI6' in g][0], self.wd_loc)
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+
+        outs = glob.glob(self.wd_loc + '/*')
+        assert(len(outs) == 4), outs
+
+    def test_parse_stb(self):
+        '''
+        test parse_stb.py basicaly
+        '''
+        script_loc = os.path.join(str(os.getcwd()),'../helper_scripts/parse_stb.py')
+        out_loc = self.wd_loc + '/test.stb'
+
+        cmd = "{0} --reverse -f {1} -o {2}".format(script_loc, ' '.join(self.genomes), out_loc)
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+
+        db = pd.read_csv(out_loc, sep='\t', names=['scaffold', 'bin'])
+        assert len(db) == 124
+        assert len(db['bin'].unique()) == 5
+
+        # Do it the regular way
+        new_out = out_loc + '_genomes_'
+        cmd = "{0} -f {1} -s {2} -o {3}".format(script_loc, self.genomes[0], out_loc, new_out)
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+
+        out_genomes = glob.glob(new_out + '*')
+        assert len(out_genomes) == 1
+
+    def test_parse_stb2(self):
+        '''
+        test parse_stb.py on zipped genomes
+        '''
+        script_loc = os.path.join(str(os.getcwd()),'../helper_scripts/parse_stb.py')
+        out_loc = self.wd_loc + '/test.stb'
+
+        cmd = "{0} --reverse -f {1} -o {2}".format(script_loc, self.zipped_genome, out_loc)
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+
+        db = pd.read_csv(out_loc, sep='\t', names=['scaffold', 'bin'])
+        assert len(db) == 1
+        print(db)
+
+        # Do it the regular way
+        new_out = out_loc + '_genomes_'
+        cmd = "{0} -f {1} -s {2} -o {3}".format(script_loc, self.zipped_genome, out_loc, new_out)
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+
+        out_genomes = glob.glob(new_out + '*')
+        assert len(out_genomes) == 1, out_genomes
 
     def tearDown(self):
         #logging.shutdown()
@@ -1725,13 +1821,13 @@ def test_unit():
     unit_test()
 
 if __name__ == '__main__':
-    rerun_test()
-    unit_test()
-    dereplicate_test()
-    filter_test()
-    cluster_test()
-    choose_test()
-    analyze_test()
+    # rerun_test()
+    # unit_test()
+    # dereplicate_test()
+    # filter_test()
+    # cluster_test()
+    # choose_test()
+    # analyze_test()
     bonus_test()
     # taxonomy_test()