+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## [2.6.1] - 2020-05-22
+- Numerous improvements to ScaffoldLevel_dRep.py, including ability to process in chunks
+- Update parse_stb.py to handle zipped .fasta files
+
 ## [2.6.0] - 2020-05-14
 - Add helper scripts ScaffoldLevel_dRep.py and parse_stb.py